modify mongo timeout
